requirements.txt
alembic==1.14.0
Flask==3.1.0
Flask_Bcrypt==1.0.1
Flask_Login==0.6.3
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
psutil==6.1.0
python-telegram-bot==21.9
SQLAlchemy==2.0.36
